Moving Like Functions into consolidated source files
